feat: 🎸 Helper methods for creating JWT
