+++ messaging/lib/index.js
@@ -1,23 +1,7 @@
-const Joi = require('joi')
 const schemas = require('./schemas')
-
-// Does not have a signature. Read more: https://tools.ietf.org/html/rfc7519#section-6
-const unsecuredMessages = [ 'CONNECTION_INFO_REQUEST' ]
-
-const validateMessage = async msg => {
-  await Joi.object({
-    type: Joi.string().valid(Object.keys(schemas)).required()
-  }).validate({ type: msg.type })
-  await schemas[msg.type].validate(msg)
-}
-
-const validateHeader = async header => {
-  await schemas.JOSE_HEADER.validate(header)
-}
+const token = require('token')
 
 module.exports = {
-  alg: schemas.alg,
-  validateMessage,
-  validateHeader,
-  unsecuredMessages
+  schemas,
+  token
 }
