+++ messaging/lib/token.js
@@ -1,8 +1,8 @@
 const axios = require('axios')
-const schemas = require('../lib/schemas')
+const schemas = require('./schemas')
 
-async function validate ({ decode, verify, parseKey }, token) {
-  if (typeof decode !== 'function' || typeof decode !== 'function' || typeof parseKey !== 'function') {
+async function verifyToken ({ decode, verify, importKey }, token) {
+  if (typeof decode !== 'function' || typeof decode !== 'function' || typeof importKey !== 'function') {
     throw new Error('First argument must be a JWT library')
   }
   const { header, payload, signature } = decode(token, { complete: true })
@@ -29,11 +29,11 @@ async function validate ({ decode, verify, parseKey }, token) {
 
   let key
   if (isDeviceIssued) {
-    key = parseKey(header.jwk)
+    key = importKey(header.jwk)
   } else {
     try {
       const { data } = await axios.get(kid)
-      key = parseKey(data)
+      key = importKey(data)
     } catch (error) {
       throw new Error(`No key found for kid: ${kid}`, error)
     }
@@ -44,4 +44,22 @@ async function validate ({ decode, verify, parseKey }, token) {
   return verify(token, key, { complete: true })
 }
 
-module.exports = validate
+async function createToken ({ sign, decode }, data, key, options) {
+  if (!data.type) {
+    throw new Error('Payload must have a type')
+  }
+  if (!schemas[data.type]) {
+    throw new Error(`Unknown schema ${data.type}`)
+  }
+  const token = await sign(data, key, options)
+  const { header, payload } = decode(token, { complete: true })
+  await schemas.JOSE_HEADER.validate(header)
+  await schemas.validate(payload)
+
+  return token
+}
+
+module.exports = ({ sign, decode, verify, importKey }) => ({
+  verify: (token) => verifyToken({ decode, verify, importKey }, token),
+  sign: (payload, key, options) => createToken({ sign, decode }, payload, key, options)
+})
